Adding Line Item Table/Model/Handlers & Client Model/Manager

+++ migrations/20160811112522_AddLineItemTable/up.sql
@@ -3,7 +3,7 @@ CREATE TABLE IF NOT EXISTS line_items (
   id serial PRIMARY KEY,
   user_id integer NOT NULL,
   credit_id integer NOT NULL,
-  amount Numeric(20, 2) NOT NULL,
+  amount float NOT NULL,
   created_date timestamp without time zone NOT NULL 
   );
 -- TODO: Add foreign key constraints
